Refactored to integrate flask_login module to products and orders views
